[NCL] Fix skia web support
